Add test and move packages under scheduled
